# Speedup install dependencies