--- a/environment/os/centos-10.sh
+++ b/environment/os/centos-10.sh
@@ -106,7 +106,7 @@ check() {
 
     # Check custom packages with bash logic
     for pkg in "${custom_packages[@]}"; do
-        missing_pkg=$(check_custom_package "$pkg")
+        missing_pkg=$(check_custom_package "$pkg" || true)
         if [ $? -eq 0 ]; then
             if [ -n "$missing_pkg" ]; then
                 missing_custom="$missing_pkg $missing_custom"
@@ -185,7 +185,7 @@ install() {
 
     # Install custom packages with bash logic
     install_custom_packages "${custom_packages[@]}"
-    
+
     # Handle non-custom packages that need special installation
     for pkg in "${custom_packages[@]}"; do
         case "$pkg" in