--- a/environment/os/centos-9.sh
+++ b/environment/os/centos-9.sh
@@ -60,8 +60,8 @@ MEMGRAPH_BUILD_DEPS=(
     libcurl-devel # mg-requests
     rpm-build rpmlint # for RPM package building
     doxygen graphviz # source documentation generators
-    which nodejs golang custom-golang1.18.9 # for driver tests
-    zip unzip java-11-openjdk-devel java-17-openjdk java-17-openjdk-devel custom-maven3.9.3 # for driver tests
+    which nodejs golang custom-golang # for driver tests
+    zip unzip java-11-openjdk-devel java-17-openjdk java-17-openjdk-devel custom-maven # for driver tests
     sbcl # for custom Lisp C++ preprocessing
     autoconf # for jemalloc code generation
     libtool  # for protobuf code generation
@@ -187,20 +187,11 @@ install() {
     fi
 
     # Install custom packages with bash logic
+    install_custom_packages "${custom_packages[@]}"
+    
+    # Handle non-custom packages that need special installation
     for pkg in "${custom_packages[@]}"; do
         case "$pkg" in
-            custom-maven3.9.3)
-                install_custom_maven "3.9.3"
-                ;;
-            custom-golang1.18.9)
-                install_custom_golang "1.18.9"
-                ;;
-            custom-rust)
-                install_rust "1.80"
-                ;;
-            custom-node)
-                install_node "20"
-                ;;
             libipt)
                 if ! dnf list installed libipt >/dev/null 2>/dev/null; then
                     dnf install -y http://repo.okay.com.mx/centos/8/x86_64/release/libipt-1.6.1-8.el8.x86_64.rpm
@@ -231,6 +222,9 @@ install() {
                     sudo -H -u "$SUDO_USER" bash -c "pip3 install virtualenv"
                     sudo -H -u "$SUDO_USER" bash -c "pip3 install virtualenvwrapper"
                 fi
+                ;;
+            *)
+                # Skip packages that don't need special handling
                 ;;
         esac
     done