--- a/release/rpm/rpmlintrc_centos10
+++ b/release/rpm/rpmlintrc_centos10
@@ -3,5 +3,5 @@
 # We are not packaging log dir
 addFilter("E: logrotate-log-dir-not-packaged")
 
-# ignore the FSF-address mismatch in libstdc++’s COPYING
+# ignore the FSF-address mismatch in libstdc++'s COPYING
 addFilter("E: incorrect-fsf-address")