… environment/os/amzn-2.sh
… environment/os/centos-7.sh
… environment/os/debian-10.sh
… environment/os/debian-12-arm.sh
… environment/os/debian-12.sh
… environment/os/debian-13-arm.sh
… environment/os/debian-13.sh
… environment/os/fedora-36.sh
… environment/os/fedora-38.sh
… environment/os/fedora-39.sh
… environment/os/fedora-41.sh
… environment/os/fedora-42.sh
… environment/os/install_deps.sh
… environment/os/output.txt
… environment/os/rocky-10.sh
… environment/os/rocky-9.sh
… environment/os/ubuntu-18.04.sh
… environment/os/ubuntu-20.04.sh
… environment/os/ubuntu-22.04-arm.sh
… environment/os/ubuntu-22.04.sh
… environment/os/ubuntu-24.04-arm.sh
… environment/os/ubuntu-24.04.sh
… environment/util.sh
… release/package/mgbuild.sh