Changes to `environment/os/` bash scripts 
- Install `memgraph` dependencies more efficiently by checking for missing packages and installing with a single command where possible, rather than looping through each package and calling `apt` or `dnf`
- Added initial scripts for `debian-13`, `fedora-42` and `rocky-10` in preparation for #2691 
- Removed scripts for unsupported distros: `amzn-2`, `centos-7`, `debian-10`, `fedora-36`, `fedora-38`, `fedora-39`, `ubuntu-18.04` and `ubuntu-20.04`


Should speed up CI a little, e.g. for Ubuntu 24.04:
```
+---------------------------------------------+-------------------------------------+-------------------------------------+
| Machine ->                                  | Matt's Desktop (Ryzen 9 3900X)      | Thanos (Xeon E5-2650)               |
+---------------------------------------------+------------------+------------------+------------------+------------------+
| Operation                                   | Before (s)       | After (s)        | Before (s)       | After (s)        |
+---------------------------------------------+------------------+------------------+------------------+------------------+
| Install Toolchain Dependencies              | 24.97  +/- 9.75  | 24.88  +/- 14.46 | 36.70  +/- 0.44  | 6.82   +/- 0.22  |
| Install Build Dependencies                  | 40.85  +/- 4.28  | 3.16   +/- 0.20  | 66.47  +/- 0.82  | 6.21   +/- 0.21  |
| Run Init Script                             | 136.33 +/- 6.6   | 123.82 +/- 4.17  | 226.15 +/- 0.79  | 224.39 +/- 2.79  |
| Total                                       | ~202             | ~151             | ~329             | 237              |
+---------------------------------------------+------------------+------------------+------------------+------------------+
```